# Conflicts:
#	server/src/main/java/org/apache/iotdb/db/conf/IoTDBConfig.java
#	server/src/main/java/org/apache/iotdb/db/conf/IoTDBDescriptor.java
#	server/src/main/java/org/apache/iotdb/db/engine/merge/inplace/task/RecoverInplaceMergeTask.java
#	server/src/main/java/org/apache/iotdb/db/engine/merge/manage/MergeResource.java
#	server/src/main/java/org/apache/iotdb/db/engine/merge/selector/MaxFileMergeFileSelector.java
#	server/src/main/java/org/apache/iotdb/db/engine/storagegroup/StorageGroupProcessor.java
#	server/src/test/java/org/apache/iotdb/db/engine/merge/MergeOverLapTest.java
#	server/src/test/java/org/apache/iotdb/db/engine/merge/MergeTest.java
#	server/src/test/java/org/apache/iotdb/db/engine/storagegroup/StorageGroupProcessorTest.java
#	tsfile/src/main/java/org/apache/iotdb/tsfile/read/common/Chunk.java
#	tsfile/src/main/java/org/apache/iotdb/tsfile/write/chunk/ChunkBuffer.java
#	tsfile/src/main/java/org/apache/iotdb/tsfile/write/chunk/ChunkWriterImpl.java
#	tsfile/src/main/java/org/apache/iotdb/tsfile/write/writer/TsFileIOWriter.java

+++ tsfile/src/main/java/org/apache/iotdb/tsfile/write/chunk/IChunkWriter.java
@@ -125,5 +125,5 @@ public interface IChunkWriter {
 
   TSDataType getDataType();
 
-  int getPtNum();
+  long getPtNum();
 }
